Implemented handling exceptions for commands of cars.

+++ GarageManager.UI/Commands/Cars/DeleteCarCommand.cs
@@ -1,5 +1,7 @@
 ﻿using GarageManager.Data.Entities;
+using GarageManager.Services.Exceptions;
 using GarageManager.Services.Interfaces;
+using GarageManager.UI.Infrastructure;
 using GarageManager.UI.ViewModels;
 using System;
 using System.Collections.Generic;
@@ -12,8 +14,9 @@ namespace GarageManager.UI.Commands.Cars
     {
         private readonly CarsListViewModel carsListViewModel;
         private readonly ICarsService carsService;
+        private readonly IMessageBoxService messageBoxService;
 
-        public DeleteCarCommand(CarsListViewModel carsListViewModel, ICarsService carsService)
+        public DeleteCarCommand(CarsListViewModel carsListViewModel, ICarsService carsService, IMessageBoxService messageBoxService)
         {
             this.carsListViewModel = carsListViewModel;
             this.carsService = carsService;
@@ -21,9 +24,25 @@ namespace GarageManager.UI.Commands.Cars
 
         public override async Task ExecuteAsync(object parameter)
         {
-            await carsService.DeleteCar(carsListViewModel.SelectedCar.CarId);
-            IEnumerable<Car> cars = await carsService.GetCars();
-            carsListViewModel.Cars = cars;
+            try
+            {
+                bool isDeleteOperationConfirmed = messageBoxService.ShowConfirmationMessageBox($"Are you sure you want to delete car {carsListViewModel.SelectedCar.Vin}?", "Delete car");
+                if(isDeleteOperationConfirmed)
+                {
+                    await carsService.DeleteCar(carsListViewModel.SelectedCar.CarId);
+                    IEnumerable<Car> cars = await carsService.GetCars();
+                    carsListViewModel.Cars = cars;
+                    messageBoxService.ShowInformationMessageBox("Car was successfully deleted.", "Delete car");
+                }
+            }
+            catch (CarNotFoundException ex)
+            {
+                messageBoxService.ShowErrorMessageBox($"Selected car with ID: {ex.CarId} not found.", "Error");
+            }
+            catch (Exception)
+            {
+                messageBoxService.ShowErrorMessageBox("Failed to delete car.", "Error");
+            }
         }
 
         public override bool CanExecute(object parameter)
